Tear down test restuls and fix syntax error
